add holy symbols
I knew from the outset I was going to want symbols made from different materials, so I followed an approach similar to the one I used with casks. However, here it seems like too much abstraction for the benefits. At least it was a good use for the gold leaf recipe.