change default unit to be u.item
Without this change, if an item weighed 7.3 lbs, Recipe.price() would return the cost for "1 lb worth" of that item even if that was nonsensical. That didn't affect the final pricing because I was always multiplying by the weight in such scenarios, but it was getting in the way of converting the item price to copper for comparing to character's cash on the new interactive trade tables. Also, I was frequently having to explicitly set `unit=1*u.item` on lots of Recipes, and everywhere that the weight of the item *was* the desired unit, I wanted to make that explicit (a pattern which was already emerging in the form of `xyz_sale_unit` variables.)